ml_system/requirements/essential.txt
# Essential ML packages for Madison Metro system
# Core ML libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Experiment tracking
wandb>=0.15.0

# API and serving
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# Data processing
pyarrow>=12.0.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
